requirements.txt
# Core PDF Processing
PyPDF2==3.0.1
pdfplumber==0.10.3
pymupdf==1.23.8
pdf2image==1.16.3

# OCR and Document Understanding
pytesseract==0.3.10
Pillow==10.1.0
opencv-python==4.8.1.78

# Machine Learning & NLP
transformers==4.35.2
torch==2.1.1
sentence-transformers==2.2.2
spacy==3.7.2
nltk==3.8.1
scikit-learn==1.3.2

# Vector Search & Similarity
faiss-cpu==1.7.4
numpy==1.24.3
scipy==1.11.4

# Text Processing
textract==1.6.5
python-docx==1.1.0
openpyxl==3.1.2

# Async Processing
asyncio==3.4.3
aiofiles==23.2.1
concurrent-futures==3.1.1

# API & Web Framework
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0

# Data Processing
pandas==2.1.4
pydantic==2.5.0
python-multipart==0.0.6

# Caching & Performance
redis==5.0.1
cachetools==5.3.2
joblib==1.3.2

# Utilities
tqdm==4.66.1
click==8.1.7
python-dotenv==1.0.0
loguru==0.7.2

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Development
black==23.11.0
flake8==6.1.0
mypy==1.7.1 